Add cubing bounds.
If the bounds passed into the constructor are not equal in width, length, and height,
the bounds will be set in such a way that the three dimensions are equal
(i.e. the bounds will be a cube). The centroid of the bounds will be preserved,
and the two smaller dimensions will be set equal to the largest dimension.