Fix a divide by 0 error.
Return 0 vector for magentic field  when a particle is overlapping with a Barnes Hut cell.
Realistically, it does not matter what the return value is,
since the Barnes Hut algorithm would calculate forces from individual particles,
rather than from the center of charge.
However, 0 seems to be the most sensible value.