Fix some problems with ticks
Fix a problem where progress was always displayed as 100%.
It attempted to check for a divide by zero and print 100%
if the number of ticks was zero, but it instead printed 100%
when the number of ticks was *not* zero.

Fix another problem where the initial state of the particles was not plotted.
Change number of ticks so that setting the simulation to 0 ticks
will only run the initial state.